# lib/api-zod/src/generated/types/crystallizerOkResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrystallizerOkResponseStatus } from "./crystallizerOkResponseStatus";
import type { ObjectiveOutput } from "./objectiveOutput";

export interface CrystallizerOkResponse {
  status: CrystallizerOkResponseStatus;
  objective: ObjectiveOutput;
}